+++ .changeset/chilled-jeans-help.md
@@ -1,0 +1,5 @@
+---
+'@api3/airnode-node': minor
+---
+
+Remove `callApi` call from `testApi` handler
